dom/tests/mochitest/chrome/window_activation.xhtml

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="data:text/css,

%23box {
  background: blue;
}

%23box:-moz-window-inactive {
  background: cyan;
}

" type="text/css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">

<box id="box" height="100"/>

<script type="application/javascript"><![CDATA[

var ok = window.arguments[0].ok;
var complete = window.arguments[0].complete;
var openerDoc = window.arguments[0].document;
var SimpleTest = window.arguments[0].SimpleTest;

SimpleTest.waitForFocus(function () {
  ok(getComputedStyle(document.getElementById("box"), "").backgroundColor, "rgb(0, 0, 255)");
  ok(getComputedStyle(openerDoc.getElementById("box"), "").backgroundColor, "rgb(0, 255, 255)");
  window.arguments[0].focus();
  ok(getComputedStyle(document.getElementById("box"), "").backgroundColor, "rgb(0, 255, 255)");
  ok(getComputedStyle(openerDoc.getElementById("box"), "").backgroundColor, "rgb(0, 0, 255)");
  complete();
}, window);

]]></script>

</window>
